[#3912] Rename `HealingActivity` to `HealActivity`
All the other activity types are named in imperative form
(e.g. "to attack", "to damage", "to summon"), so this updates
heal to match.
