+++ lang/en.json
@@ -1419,8 +1419,8 @@
 "DND5E.Hair": "Hair",
 "DND5E.HalfProficient": "Half Proficient",
 
-"DND5E.HEALING": {
-  "Title": "Healing",
+"DND5E.HEAL": {
+  "Title": "Heal",
   "FIELDS": {
     "healing": {
       "label": "Healing",
