+++ module/applications/activity/_module.mjs
@@ -2,7 +2,7 @@ export {default as ActivitySheet} from "./activity-sheet.mjs";
 export {default as AttackSheet} from "./attack-sheet.mjs";
 export {default as DamageSheet} from "./damage-sheet.mjs";
 export {default as EnchantSheet} from "./enchant-sheet.mjs";
-export {default as HealingSheet} from "./healing-sheet.mjs";
+export {default as HealSheet} from "./heal-sheet.mjs";
 export {default as SummonSheet} from "./summon-sheet.mjs";
 export {default as UtilitySheet} from "./utility-sheet.mjs";
