+++ module/data/activity/_module.mjs
@@ -3,7 +3,7 @@ export {default as BaseActivityData} from "./base-activity.mjs";
 export {default as AttackActivityData} from "./attack-data.mjs";
 export {default as DamageActivityData} from "./damage-data.mjs";
 export {default as EnchantActivityData} from "./enchant-data.mjs";
-export {default as HealingActivityData} from "./healing-data.mjs";
+export {default as HealActivityData} from "./heal-data.mjs";
 export {default as SaveActivityData} from "./save-data.mjs";
 export {default as SummonActivityData} from "./summon-data.mjs";
 export {default as UtilityActivityData} from "./utility-data.mjs";
