+++ module/data/activity/heal-data.mjs
@@ -2,11 +2,11 @@ import DamageField from "../shared/damage-field.mjs";
 import BaseActivityData from "./base-activity.mjs";
 
 /**
- * Data model for an healing activity.
+ * Data model for an heal activity.
  *
  * @property {DamageData} healing
  */
-export default class HealingActivityData extends BaseActivityData {
+export default class HealActivityData extends BaseActivityData {
   /** @inheritDoc */
   static defineSchema() {
     return {
