+++ module/documents/activity/_module.mjs
@@ -3,7 +3,7 @@ export {default as ActivityMixin} from "./mixin.mjs";
 export {default as AttackActivity} from "./attack.mjs";
 export {default as DamageActivity} from "./damage.mjs";
 export {default as EnchantActivity} from "./enchant.mjs";
-export {default as HealingActivity} from "./healing.mjs";
+export {default as HealActivity} from "./heal.mjs";
 export {default as SaveActivity} from "./save.mjs";
 export {default as SummonActivity} from "./summon.mjs";
 export {default as UtilityActivity} from "./utility.mjs";
