+++ templates/activity/heal-effect.hbs
@@ -1,4 +1,4 @@
 <section class="tab activity-{{ tab.id }} {{ tab.cssClass }}" data-tab="{{ tab.id }}" data-group="{{ tab.group }}">
-    {{> "systems/dnd5e/templates/activity/parts/healing-damage.hbs" }}
+    {{> "systems/dnd5e/templates/activity/parts/heal-healing.hbs" }}
     {{> "systems/dnd5e/templates/activity/parts/activity-effects.hbs" }}
 </section>
